+++ src/main/java/com/revature/_1811_nov27_wvu/icebox/entity/File.java
@@ -1,4 +1,4 @@
-package com.revature.beans;
+package com.revature._1811_nov27_wvu.icebox.entity;
 
 import java.sql.Date;
 import java.util.Set;
